[SCRUM-45] refactor: More menu 공통 컴포넌트화 및 Tooltip 추가
- More menu를 shared 컴포넌트로 이동하여 재사용성 향상
- Shadcn Tooltip 컴포넌트 적용
- 더보기 버튼 hover 시 툴팁 표시
- 컴포넌트 JSDoc 문서화 추가
